operis-user/operis-user-service/src/main/java/com/operis/service/UserService.java
package com.operis.service;

import com.operis.dto.UserDTO;
import com.operis.mapper.UserMapper;
import com.operis.model.User;
import com.operis.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private UserMapper userMapper;

    public UserDTO createUser(UserDTO userDTO) {
        User user = userMapper.toUser(userDTO);
        user = userRepository.save(user);
        return userMapper.toUserDTO(user);
    }
}
